Retrieve paths to longer strings Returns paths to (markdown) files containing longer string parameters.

Retrieve paths to longer strings Returns paths to (markdown) files containing longer string parameters.

strings_from_mdfile(filenames = NULL, path)

Arguments

filenames
[character()] giving the filenames of longer strings expected at path, without extension. Defaults to NULL, in which case all *.mds at path are returned.

path
[character(1)] giving the path to a directory including *.mds with longer strings.

Value

[character()] giving absolute paths to files, named by filenames without extension.

Storing longer strings

Longer strings, such as for survey questions or user interface are often best edited in separate files with proper markup (markdown) support. In the context of a package, they are best saved ininst/, which can be conveniently accessed with system.file().